Link to main site on first mention of 'LRUG'
I noticed that we don't really link to the homepage in this README, so this commit makes the first mention of 'LRUG' a link to the homepage.

This also acts as an acronym definition (doing HTML abbreviations is a markdown extension that probably isn't supported), seeing as we don't define the acronym in this document either. It is helpfully at the top of the LRUG homepage though.