Make some tweaks to video editing
Add images to show the desired timeline for the intro + outro's and clarify the text.  This should make it clearer what we're trying to achieve.

Change the information about editing the audio.  We get better results with "voice enhance" and a lower volume boost vs. a pure 400% volume boost on it's own.