De-twitter-and-lanyrd-ify the "keep in touch" section
Lanyrd is dead, and has been for a while.  Twitter is dying and gross.  We replace the references with eventbrite (for now at least) and mastodon respectively.  We also reorder the options to fall in the order we'd prefer people to find out about LRUG events: mailing-list, website, calendar, mastodon, eventbrite.