De-Skills-Matter-ify the readme
It's only taken 4 years, but here we are.  A version of the readme that doesn't talk about skills matter.  This mostly means tidying up the after the talks and sponsorship sections - we don't go to The Singer Tavern, and we do need a venue sponsor.  There are also tweaks to the Talks section to stop talking about archaic projectors.  Finally some generic tweaks throughout to stop talking about Skills Matter and provide more general advice on how LRUG runs.