Update link to "how to be a good ally" to https
We got reports of this link sometimes redirecting to a scammy bot-detection page.  We verified we sometimes were seeing this and that changing to https did seem to help.  The content is good so we don't want to lose the link, but we also don't want to point people at scams.